+++ src/ingest_validation_tools/plugin_validator.py
@@ -43,15 +43,25 @@ class Validator(object):
     """float: a rough measure of cost to run.  Lower is better.
     """
 
-    def __init__(self, base_path: PathOrStr, assay_type: str, contains: List = []):
+    def __init__(
+        self, base_paths: List[Path], assay_type: str, contains: List = [], **kwargs
+    ):
         """
-        base_path is expected to be a directory.
-        This is the root path of the directory tree to be validated.
+        base_paths is expected to be a list of directories.
+        These are the root paths of the directory trees to be validated.
         assay_type is an assay type, one of a known set of strings.
         """
-        self.path = Path(base_path)
-        if not self.path.is_dir():
-            raise ValidatorError(f"{self.path} is not a directory")
+        del kwargs
+        if isinstance(base_paths, List):
+            self.paths = [Path(path) for path in base_paths]
+        elif isinstance(base_paths, Path):
+            self.paths = [base_paths]
+        elif isinstance(base_paths, str):
+            self.paths = [Path(base_paths)]
+        else:
+            raise Exception(
+                f"Validator init received base_paths arg as type {type(base_paths)}"
+            )
         self.assay_type = assay_type
         self.contains = contains
 
@@ -88,15 +98,21 @@ def run_plugin_validators_iter(
                     f"{metadata_path} contains more than one assay type"
                 )
 
+    data_paths = []
     if all("data_path" in row for row in sv.rows):
         for row in sv.rows:
             data_path = Path(row["data_path"])
             if not data_path.is_absolute():
                 data_path = (Path(metadata_path).parent / data_path).resolve()
-            for k, v in validation_error_iter(
-                data_path, sv.dataset_type, plugin_dir, sv.contains, **kwargs
-            ):
-                yield k, v
+            if not data_path.is_dir():
+                raise ValidatorError(
+                    f"{data_path} should be the base directory of a dataset"
+                )
+            data_paths.append(data_path)
+        for k, v in validation_error_iter(
+            data_paths, sv.dataset_type, plugin_dir, sv.contains, **kwargs
+        ):
+            yield k, v
     else:
         raise ValidatorError(f"{metadata_path} is missing values in 'data_path' column")
 
@@ -138,18 +154,18 @@ def validation_class_iter(plugin_dir: PathOrStr) -> Iterator[Type[Validator]]:
 
 
 def validation_error_iter(
-    base_dir: PathOrStr,
+    base_dir: List[Path],
     assay_type: str,
     plugin_dir: PathOrStr,
     contains: List,
     **kwargs,
 ) -> Iterator[KeyValuePair]:
     """
-    Given a base directory pointing to a tree of upload data files and a
-    path to a directory of Validator plugins, iterate over the results of
-    applying all the plugin validators to the directory tree.
+    Given a list of base directories in the upload and a path to a directory
+    of Validator plugins, iterate over the results of applying all the plugin
+    validators to each directory tree.
 
-    base_dir: the root of a directory tree of upload data files
+    base_dir: a list of paths representing the datasets in an upload
     assay_type: the assay type which produced the data in the directory tree
     plugin_dir: path to a directory containing classes derived from Validator
     contains: list of component assay types (empty if not multi-assay)
@@ -157,9 +173,6 @@ def validation_error_iter(
     returns an iterator the values of which are key value pairs representing
     error messages
     """
-    base_dir = Path(base_dir)
-    if not base_dir.is_dir():
-        raise ValidatorError(f"{base_dir} should be the base directory of a dataset")
     for cls in validation_class_iter(plugin_dir):
         validator = cls(base_dir, assay_type, contains)
         for err in validator.collect_errors(**kwargs):  # type: ignore